{"schema":"decodex/commit/1","summary":"Split stale active worktree inspection","authority":"XY-1115"}
